+++ src/internal/THEOplayerView.tsx
@@ -50,7 +50,7 @@ import type {
   NativeTextTrackListEvent,
 } from './adapter/event/native/NativeTrackEvent';
 import { toMediaTrackType, toMediaTrackTypeEventType, toTextTrackEventType, toTrackListEventType } from './adapter/event/native/NativeTrackEvent';
-import type {
+import {
   NativeDurationChangeEvent,
   NativeErrorEvent,
   NativeLoadedMetadataEvent,
@@ -107,6 +107,7 @@ interface THEOplayerRCTViewProps {
   onNativeTHEOliveEvent: (event: NativeSyntheticEvent<NativeTheoLiveEvent>) => void;
   onNativeCastEvent: (event: NativeSyntheticEvent<NativeCastEvent>) => void;
   onNativePresentationModeChange: (event: NativeSyntheticEvent<NativePresentationModeChangeEvent>) => void;
+  onNativeDeviceOrientationChanged: () => void;
   onNativeResize: (event: NativeSyntheticEvent<NativeResizeEvent>) => void;
 }
 
@@ -141,7 +142,10 @@ export class THEOplayerView extends PureComponent<React.PropsWithChildren<THEOpl
   }
 
   componentDidMount() {
-    this._dimensionsHandler = Dimensions.addEventListener('change', this._onDimensionsChanged);
+    if (Platform.OS !== 'ios') {
+      // On iOS we use the native deviceOrientation event.
+      this._dimensionsHandler = Dimensions.addEventListener('change', this._onDimensionsChanged);
+    }
   }
 
   componentWillUnmount() {
@@ -169,6 +173,14 @@ export class THEOplayerView extends PureComponent<React.PropsWithChildren<THEOpl
 
   private _onDimensionsChanged = () => {
     this.setState({ screenSize: getFullscreenSize() });
+  };
+
+  private _onDeviceOrientationChanged = () => {
+    if (Platform.OS === 'ios') {
+      // On iOS, we use the native deviceOrientation event to update the screenSize
+      // because of an issue on iPad with React-native's Dimensions.
+      this._onDimensionsChanged();
+    }
   };
 
   private _onNativePlayerReady = (event: NativeSyntheticEvent<NativePlayerStateEvent>) => {
@@ -424,6 +436,7 @@ export class THEOplayerView extends PureComponent<React.PropsWithChildren<THEOpl
           onNativeTHEOliveEvent={this._onTHEOliveEvent}
           onNativeCastEvent={this._onCastEvent}
           onNativePresentationModeChange={this._onPresentationModeChange}
+          onNativeDeviceOrientationChanged={this._onDeviceOrientationChanged}
           onNativeResize={this._onResize}
         />
         {posterActive && <Poster uri={poster} style={posterStyle} />}